+++ src/main/java/com/example/yztx/repository/UserRepository.java
@@ -7,10 +7,10 @@ public interface UserRepository extends CrudRepository<User,Long> {
     //账号登录使用
     User findUserByAccountAndPassword(String account,String password);
     //手机号登录使用
-    User findUserByPhoneAndPassword(int phone,String password);
+    User findUserByPhoneAndPassword(long phone,String password);
     //查询账号是否已存在
     Boolean existsUserByAccount(String account);
     //查询手机号、账号是否已被绑定
-    User findUserByPhone(int phone);
+    User findUserByPhone(long phone);
     User findUserByAccount(String account);
 }
